Meta: Added a batch script that generates the build files
In order to streamline the process of invoking CMake in order to
generate the files that are used by build systems such as Ninja or
Visual Studio a batch file can now be used.

Previously, CMake had to be manually invoked from the command
line, so making subtle errors was a lot more common.