AT: Removed 'using namespace AT' from the global namespace
Because the AT framework is used everywhere in the Moonrise project,
it's not ergonomic to write AT:: everytime you want to access
something from the AT library.

We used to include the AT namespace into the global namespace. That
method is far from ideal, so now every class/struct/function has to
be manually included in the global namespace.