AT: Implemented ErrorOr as a method to propagate errors
ErrorOr is a templated class that stores an error object or a value,
as well as a flag that tells which one of them. In composition with
the TRY and MUST macros, it represents the foundation of error
propagation in the call-stack.

Its purpose is to provide a common and simple method of handling
errors, that is consistent across the entire Moonrise project. By
implementing features that make error propagation very easy, even
though it is not the most performant method, we hope that the
robustness of our system will greatly improve. :)

It is inspired by the Serenity AK type with the same name and the
Rust programming language method of handling errors.