AT: Reworked the String and StringView API for better consistency
The API for creating strings is now consistent across the String and
StringView classes, while also leveraging the benefits of the ErrorOr
return type.

The API for accessing the raw bytes from a string object has been
significantly simplified - a single type of member functions can now
be used to get a span that contains the bytes, and any other method
of accessing the bytes was removed! ;)