AT: Fixed bug that would try to free an incomplete string container
In the old String::create() implementation we would create the string
object at the top of the function, as we may need its inline buffer
for the small string optimization. However, there is a hidden early
return in TRY_ASSIGN, in which case the local string will be
destroyed when it is in an incomplete form.

The fix was to create the string object after the heap memory was
successfully allocated and no other failure points exist.