AT: Implemented the API and core architecture of string formatting
Because we wanted string formatting to be both type-safe and easy to
use, we decided to not use C-style string formatting, nor the C++
approach. Instead, we heavily inspired our API from the C# syntax.

Currently, string formatting is performed with a single format()
function, that returns a new string object that holds the formatted
result. In the future, we expect a similar function that will store
the result to a buffer provided by the user instead.

Formatting of the arguments happens in the matching specialization of:

struct Formatter<T>::format(FormatBuilder&,
                            const FormatBuilder::Specifier&,
                            const T&);

The format builder manages the buffer of the formatted string, as well
as parsing the string to format. Specializations of the function above
are essentially a composition of the builder built-in member functions.