Core: Implemented primitive logging functions
Currently, all logs are printed to the console by calling 'printf'. In
future implementations, the output channels will be separated based on
severity and the text will be colored accordingly.

While this is a very simple implementation, it should work just fine for
an application that focuses on the GUI experience! ;)