AT: Implemented a synchronous reference counted smart pointer
The RefPtr is a templated container that stores and manages an instance
of a type derived from RefCounted<T>, using automatic reference
counting. When the reference count hits 0, the object is destroyed and
its memory is released.

The manipulation of the reference count is *not* performed atomically.
In the future, there will be an implementation for an AtomicRefCounted
type that will be thread-safe. ^^