AT: Cleaned the Vector container from the ErrorOr architecture
Moving forward, we want to restrict the usage of the ErrorOr
architecture only in systems where errors are not actually fatal. There
is no point in handling out-of-memory errors in an application designed
for graphics production - there is nothing we can do anyway!

Thus, the Vector API has been streamlined and now doesn't use any of
the ErrorOr architecture! ^^